testing if it compiles in tdaq

+++ src/main/java/ch/cern/tdaq/operator/runcontroller/CustomResource/RunControllerCRResourceSpec.java
@@ -10,12 +10,12 @@ import io.fabric8.kubernetes.api.model.KubernetesResource;
 )
 public class RunControllerCRResourceSpec implements KubernetesResource {
     private String name;
-    private int runNumber;
+    private long runNumber;
     private String runPipe;
     private String label;
 
-    public int getRunNumber() { return runNumber; }
-    public void setRunNumber(int runNumber) { this.runNumber = runNumber; }
+    public long getRunNumber() { return runNumber; }
+    public void setRunNumber(long runNumber) { this.runNumber = runNumber; }
 
     public String getRunPipe() { return runPipe; }
     public void setRunPipe(String runPipe) { this.runPipe = runPipe; }
